[generate] Remove unlisted cases that are duplicates of generated cases
Fix #122

From now on, all listed directories and cases are generated first,
followed by all manual cases. For each generated listed case, we store
the testdate (.in file) in a dictionary. When an unlisted (manual) case
has the same input, it is assumed that the manual case used to be
generated but a renaming or renumbering occured. Thus, we check that all
files in the manual case match the corresponding files in the generated
case, and then delete the manual case if they match indeed.

If not all files match, a warning is raised and the manual case is preserved.