Fix #149: Remove indent from `ruamel.yaml` written `problems.yaml`
The reason why ruamel.yaml adds two extra spaces at the start of each
line, is the indent config (sequence=4, offset=2).
Either this indent config should be made inconsistent with the indent
configs used on other places, or we post-process the yaml to remove the
leading spaces (which I've done in this commit).