Fix #177: `config.TEST_TLE_SUBMISSIONS` bug
Before 73ff54bc83, the conditional was always `false` for TLE
submissions outside the testing scenario. After 73ff54bc83, this
conditional was always `true` instead. To fix this more properly, I've
added a dependency on the `config.RUNNING_TEST` flag to fix this. So,
submissions are now only _skipped_ (using `continue`) when:

- They should have a TLE verdict
- We are running tests (`config.RUNNING_TEST`)
- We do NOT want to test TLE sbumissions (`not config.TEST_TLE_SUBMISSIONS`)