[export] Remove TODO about reusing problem IDs because it's not possible
We could probably get away with it if DOMjudge would always use the
"Data source: configuration data external" setting.
In this mode, the problem IDs exactly match the problem shortnames.
However, when "Data source: all local", the problem IDs in the API
are numeric, and therefore cannot be matched directly with our problems.