[latex] Print paths to samples with .as_posix()
This fixes a LaTeX compile error on Windows. Because Windows uses
backslash as path separator, LaTeX thought that these paths were
commands. As a fix, we can force the paths to print as_posix() (i.e.,
with forward slashes), because Windows also accepts this.

Thanks to @alexbolfa for reporting this bug and testing this patch on
their Windows machine 😄