validate -v: show validator flags
When no validator flags are used, nothing new happens.
Thus, in BAPC tradition, the only change is that the strict flags
for the output validator when used for answer validation are made explicit, like this:

 Invalidation: invalid_answers/capitals default_output_validator.cpp (ans) [case_sensitive, space_change_sensitive]: accepted

In general, if validator flags are provided (via testdata.yaml), these
are shown in square brackets after the validator name.