[problem] Update parsing of problem.yaml based on Kattis/problem-package-format#372 (#437)
* [problem] Update parsing of problem.yaml based on Kattis/problem-package-format#372

Changes:
- **C1**: Keywords are now a list of strings (we used to still parse them as a single string, woops)
- **C2**: No change needed, this was just a "bug" in the human-readable text of the specification
- **C3**: Add constraints to float/int types in `limits` and warnings when any of the values are out of range
- **C5**: Do not allow lists to be empty (if a list-field is optional, it should be either `None` or a non-empty list)

The discussion for **C4** was moved to Kattis/problem-package-format#378 and is pending consensus, and the proposals for **C6**, **Q1**, and **Q2** were dropped.

* [problem] Fix parsing of ProblemSource, thanks to Thore's extra tests

* [test] Add some more tests for license/rights_owner

Note that I haven't thoroughly tested the combination of `license` and `rights_owner`.
Similar to embargo_until, BAPCtools doesn't really do much with this information anyway,
so the parser there is currently quite lenient, and as such I'll consider it out-of-scope for this PR.
If others feel like improving and adding tests for this, feel free to do so 🙂