[problem] get_test_case_yaml: Improve consistency of `path` argument
Previously, it could be either the path to an input file or the path to a test group directory.
However, it was assumed that it is always an input file, so `.with_suffix("yaml")` was always performed.
Now, the .yaml suffix is added outside this function, and this function expects
either the path to a <test_case>.yaml file, or the path to a test group directory.