[latex] Fix calculation of \idealwidth
Apparently, at the top-level, the \textwidth is not set to the actual paper width.
Fixed this by moving it inside \Sample, which means it is executed for every sample, but that's not too expensive.
Used some debug printing: \the\PS@inputwidth{} \the\PS@outputwidth{} \the\PS@idealwidth{} \the\textwidth{} \the\arraycolsep{}